@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,8 +27,8 @@ dependencies:
   cupertino_icons: ^1.0.2
   
   camera_camera: ^2.0.1
-  image_picker: ^0.7.4
-  get:
+  image_picker: ^0.8.1+4
+  get: ^4.1.4
 
 dev_dependencies:
   flutter_test:
