@@ -6,7 +6,7 @@ import 'package:get/get.dart';
 class PreviewPage extends StatelessWidget {
   File file;
 
-  PreviewPage({Key key, this.file}) : super(key: key);
+  PreviewPage({Key? key, required this.file}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
