@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 class Anexo extends StatelessWidget {
   final File arquivo;
 
-  Anexo({Key key, this.arquivo}) : super(key: key);
+  Anexo({Key? key, required this.arquivo}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
